+++ Source/Widgets/Plotting/AnalysisSupport/ORXYRoi.m
@@ -66,9 +66,9 @@
 		[dataSource plotter:aPlot index:i x:&x y:&y];
 		if(x>=xStart && x<xEnd){
 			sumY	+= y;
-			sumXY	+= x*y;
-			sumX2Y	+= x*x*y;
-			sumValX += y*x;
+			sumXY	+= (double)x*y;
+			sumX2Y	+= (double)x*x*y;
+			sumValX += (double)y*x;
 
 			if (y < minY) minY = y;
 			if (y > maxY) {
@@ -80,9 +80,9 @@
 	
 	
 	if(totalNum){
-		double theXAverage = sumXY / (double)sumY;
+		double theXAverage = sumXY / sumY;
 		average = theXAverage;
-		sigma	= sqrt((sumX2Y/(double)sumY) - (theXAverage*theXAverage));
+		sigma	= sqrt((sumX2Y/sumY) - (theXAverage*theXAverage));
 	}
 	else {
 		average = 0;
